Server Selector Service integrated
Now you can invoke the function autoServerSelection() inside
ServerCommunicator and this will choose the server to connnect to based
on that asp.net service I have built. If it cant find a service it will
set the defaults to localhost.

If you create a ServerCommunicator object with the default constructor
autoServerSelection will be called as part of it, therefore we should
always create ServerConnector objects using the default constructor